feat(client/customHook): add useAuth for logic by authentication
