--- client/src/methodsWallet/authAnchor/authenticationAnchor.jsx
@@ -1,12 +1,0 @@
-import { readChallenge } from "./readChallenge";
-import { signChallenge } from "./signChallenge";
-import { postChallenge } from "./postChallenge";
-
-const authenticationWithAnchor = async () => {
-  
-  
-  
-    return;
-};
-
-export default authenticationWithAnchor;
